requirements.txt
boto3
botocore
PyGithub
